Dialogs changed to angular alerts + validation and UI fixes

+++ eshop-zone/src/app/merchant/product-list/product-list.component.ts
@@ -1,8 +1,10 @@
 import { Component,OnInit} from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Product } from 'src/app/models/product';
 import { CatalogueService } from 'src/app/services/catalogue.service';
 import { NavBarService } from 'src/app/services/nav-bar.service';
 import { ProductService } from 'src/app/services/product.service';
+import { ConfirmationDialogComponent } from 'src/app/shared/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-product-list',
@@ -16,6 +18,7 @@ export class ProductListComponent implements OnInit {
     private catalogueService: CatalogueService,
     private navBarService:NavBarService,
     private productService:ProductService,
+    public dialog: MatDialog
   ) { }
 
   
@@ -30,12 +33,25 @@ export class ProductListComponent implements OnInit {
 
 
   delete(id:string,index:number,name:string){
-    if(confirm("Are you sure to delete "+name)) {
-      this.productService.deleteProduct(id).subscribe(result=> {
-        console.log(result);
-        this.getAllProducts();
-      });
-    }
+
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: {
+        title: 'Confirm',
+        content:`Are you sure to delete ${name}.`
+      }
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if(!result){
+        return;
+      }
+      if(confirm("Are you sure to delete "+name)) {
+        this.productService.deleteProduct(id).subscribe(result=> {
+          console.log(result);
+          this.getAllProducts();
+        });
+      }
+    });
 
   }
